move LogHandlerAccessor to shared so that it can be used by both tests and the launcher
